requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1
django-countries==7.3.2
django-jazzmin==2.5.0
django-phonenumber-field==6.4.0
djangorestframework==3.13.1
phonenumbers==8.12.54
Pillow==9.2.0
pytz==2022.2.1
scrapy-rotating-proxies==0.6.2
six==1.16.0
sqlparse==0.4.2
typing==3.7.4.3
typing_extensions==4.3.0
